Reenabled android tests and moved ExtendedAsynchronousByteChannel
Moved ExtendedAsynchronousByteChannel to com.mongodb.internal.connection
The dex'ing process didn't like it being referenced in a higher package
when the code was split across dex files.

JAVA-3142
