+++ driver-core/src/main/com/mongodb/internal/connection/AsynchronousChannelStream.java
@@ -26,7 +26,6 @@ import com.mongodb.connection.AsyncCompletionHandler;
 import com.mongodb.connection.BufferProvider;
 import com.mongodb.connection.SocketSettings;
 import com.mongodb.connection.Stream;
-import com.mongodb.internal.connection.tlschannel.async.ExtendedAsynchronousByteChannel;
 import org.bson.ByteBuf;
 
 import java.io.IOException;
